+++ python/GafferSceneUI/CatalogueSelectUI.py
@@ -1,6 +1,6 @@
 ##########################################################################
 #
-#  Copyright (c) 2012, John Haddon. All rights reserved.
+#  Copyright (c) 2017, Image Engine Design Inc. All rights reserved.
 #
 #  Redistribution and use in source and binary forms, with or without
 #  modification, are permitted provided that the following conditions are
@@ -33,46 +33,59 @@
 #  SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 #
 ##########################################################################
-
-import threading
-
+import IECore
 import Gaffer
-import GafferUI
+import GafferScene
 
-import GafferImage
+def __imageNames( plug ) :
+	node = plug.node()
+	imagePlug = node["in"]
 
-__all__ = []
+	while imagePlug is not None and not isinstance( imagePlug.node(), GafferScene.Catalogue ) :
+		imagePlug = imagePlug.getInput()
+
+	if imagePlug is None :
+		return []
+
+	return imagePlug.node()["images"].keys()
+
+def __imagePresetNames( plug ) :
+
+	return IECore.StringVectorData(
+		[ "Selected", "Output/1", "Output/2", "Output/3", "Output/4" ] +
+		[ "Image/" + i for i in __imageNames( plug ) ]
+	)
+
+def __imagePresetValues( plug ) :
+
+	return IECore.StringVectorData(
+		[ "", "output:1", "output:2", "output:3", "output:4" ] +
+		__imageNames( plug )
+	)
 
 Gaffer.Metadata.registerNode(
 
-	GafferImage.Display,
+	GafferScene.CatalogueSelect,
 
 	"description",
-	"""
-	Interactively displays images as they are rendered.
-
-	This node runs a server on a background thread,
-	allowing it to receive images from both local and
-	remote render processes. To set up a render to
-	output to the Display node, use an Outputs node with
-	an Interactive output configured to render to the
-	same port as is specified on the Display node.
-	""",
+	"Finds an image in a directly connected Catalogue by name.",
 
 	plugs = {
 
-		"port" : [
+		"imageName" : [
 
 			"description",
-			"""
-			The port number on which to run the display server.
-			Outputs which specify this port number will appear
-			in this node - use multiple nodes with different
-			port numbers to receive multiple images at once.
-			""",
+			"The name of the image to extract.",
 
-		],
+			"presetNames", __imagePresetNames,
+			"presetValues", __imagePresetValues,
+			# Don't promote presets so they are still computed dynamically for
+			# the promoted plug rather than being baked.
+			"presetNames:promotable", False,
+			"presetValues:promotable", False,
+			"presetsPlugValueWidget:allowCustom", True,
+			"plugValueWidget:type", "GafferUI.PresetsPlugValueWidget",
 
+		]
 	}
-
 )